+++ market-make/database/migrations/2026_07_11_000001_add_fundamentals_to_companies_table.php
@@ -9,8 +9,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('companies', function (Blueprint $table) {
-            $table->float('revenue_growth')->nullable();      // G: annualized 5Y revenue growth, %
-            $table->float('eps_growth')->nullable();          // annualized 5Y diluted EPS growth, %
+            $table->double('revenue_growth')->nullable();     // G: annualized 5Y revenue growth, %
+            $table->double('eps_growth')->nullable();         // annualized 5Y diluted EPS growth, %
             $table->integer('reliability_score')->nullable(); // R: passed checks
             $table->integer('reliability_max')->nullable();   // R: total checks (6)
             $table->text('reliability_checks')->nullable();   // R: per-check details (JSON)
